Canononicalize system containers as type: `system`
Currently for system containers, in `images list --json`, we have:
`"type": "System"`, and in `ps --json`, it shows up as "type": "systemcontainer"

The extra `"container"` suffixing the type seems redundant, so let's
kill it, and canonicalize on lowercase, hence both cases simply become
`"system`".

This matters because I am parsing the JSON in Ansible and consistency
is good.